Fix `available-at` breaking semantics of excludes and transitive=false
Whenever a variant is declared using an `available-at` redirection, we
internally use a dependency to lazily fetch the variant. This behavior
wasn't intentional but has been used in the wild, including in development
process where it's possible to "partially" publish a component, in
particular in environments like native where it's not possible to
build all variants of a component on a single machine.

This commit reworks how "available-at" is implemented so that we keep
the laziness, keep and intermediate "dependency" which can be used
to figure out that a variant is delegating to an external component,
but in this case makes the semantics of "exclusions" and "transitive=false"
consistent.