Introduce getExternalVariant API on resolution result
This API lets the user know whether a variant is just a bridge
to an externally available variant or not. In practice, an
external variant is just a variant which is "available at" another
module coordinates.

It is implemented as a dependency to the owner of the external
variant. In order to make the difference between a regular variant
and a variant which just "redirects" to an external component,
this API provides a `getExternalVariant` method which would reference
the actual _resolved_ variant from the external module.

To make it more convenient to figure out what the owner of the
external variant is, the `ResolvedVariantResult` now contains
a reference to the owner.